refactor: updating name for TestPassed
